requirements.txt
python-consul>=0.7.0
